+++ src/main/java/com/coffesoft/financeapplication/repository/monobank/MccRepository.java
@@ -3,5 +3,5 @@ package com.coffesoft.financeapplication.repository.monobank;
 import com.coffesoft.financeapplication.model.monobank.Mcc;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MccRepository extends JpaRepository<Mcc, Integer> {
+public interface MccRepository extends JpaRepository<Mcc, Long> {
 }
